pyproject.toml
[project]
name = "cp219-project-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "plotly>=5.3.0",
    "wandb>=0.12.0",
    "pyyaml>=5.4.0",
    "joblib>=1.0.0",
    "tqdm>=4.62.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "black>=21.0",
    "flake8>=3.9.0",
    "ipython>=7.25.0",
    "jupyter>=1.0.0",
]

advanced = [
    "statsmodels>=0.13.0",
    "networkx>=2.6.0",
    "xgboost>=1.5.0",
]

